chore: remove ai graph modeling
