+++ packages/db/prisma/sqlite-baseline.sql
@@ -398,113 +398,6 @@ CREATE TABLE "TenantAiSettings" (
     CONSTRAINT "TenantAiSettings_tenantId_fkey" FOREIGN KEY ("tenantId") REFERENCES "Tenant" ("id") ON DELETE CASCADE ON UPDATE CASCADE
 );
 
--- CreateTable
-CREATE TABLE "SchoolEntity" (
-    "id" TEXT NOT NULL PRIMARY KEY,
-    "tenantId" TEXT NOT NULL,
-    "type" TEXT NOT NULL,
-    "name" TEXT NOT NULL,
-    "aliases" JSONB NOT NULL DEFAULT '[]',
-    "confidence" REAL NOT NULL DEFAULT 0.5,
-    "source" TEXT NOT NULL,
-    "evidence" JSONB NOT NULL DEFAULT '[]',
-    "firstSeenAt" DATETIME NOT NULL DEFAULT CURRENT_TIMESTAMP,
-    "lastSeenAt" DATETIME NOT NULL DEFAULT CURRENT_TIMESTAMP,
-    "updatedAt" DATETIME NOT NULL,
-    CONSTRAINT "SchoolEntity_tenantId_fkey" FOREIGN KEY ("tenantId") REFERENCES "Tenant" ("id") ON DELETE CASCADE ON UPDATE CASCADE
-);
-
--- CreateTable
-CREATE TABLE "SchoolModelSnapshot" (
-    "id" TEXT NOT NULL PRIMARY KEY,
-    "tenantId" TEXT NOT NULL,
-    "version" INTEGER NOT NULL,
-    "status" TEXT NOT NULL DEFAULT 'active',
-    "summary" TEXT NOT NULL,
-    "entities" JSONB NOT NULL,
-    "modelingMemory" JSONB NOT NULL,
-    "rules" JSONB NOT NULL,
-    "metrics" JSONB NOT NULL DEFAULT '{}',
-    "createdAt" DATETIME NOT NULL DEFAULT CURRENT_TIMESTAMP,
-    CONSTRAINT "SchoolModelSnapshot_tenantId_fkey" FOREIGN KEY ("tenantId") REFERENCES "Tenant" ("id") ON DELETE CASCADE ON UPDATE CASCADE
-);
-
--- CreateTable
-CREATE TABLE "PostAiAnalysis" (
-    "id" TEXT NOT NULL PRIMARY KEY,
-    "tenantId" TEXT NOT NULL,
-    "postId" TEXT NOT NULL,
-    "modelSnapshotId" TEXT,
-    "provider" TEXT NOT NULL,
-    "model" TEXT NOT NULL,
-    "status" TEXT NOT NULL DEFAULT 'pending',
-    "confidence" REAL,
-    "categories" JSONB NOT NULL DEFAULT '[]',
-    "entities" JSONB NOT NULL DEFAULT '[]',
-    "reasons" JSONB NOT NULL DEFAULT '[]',
-    "rawOutput" JSONB,
-    "error" TEXT,
-    "createdAt" DATETIME NOT NULL DEFAULT CURRENT_TIMESTAMP,
-    "updatedAt" DATETIME NOT NULL,
-    CONSTRAINT "PostAiAnalysis_tenantId_fkey" FOREIGN KEY ("tenantId") REFERENCES "Tenant" ("id") ON DELETE CASCADE ON UPDATE CASCADE,
-    CONSTRAINT "PostAiAnalysis_postId_fkey" FOREIGN KEY ("postId") REFERENCES "Post" ("id") ON DELETE CASCADE ON UPDATE CASCADE,
-    CONSTRAINT "PostAiAnalysis_modelSnapshotId_fkey" FOREIGN KEY ("modelSnapshotId") REFERENCES "SchoolModelSnapshot" ("id") ON DELETE SET NULL ON UPDATE CASCADE
-);
-
--- CreateTable
-CREATE TABLE "AiBackfillBatch" (
-    "id" TEXT NOT NULL PRIMARY KEY,
-    "tenantId" TEXT NOT NULL,
-    "actorId" TEXT,
-    "status" TEXT NOT NULL DEFAULT 'queued',
-    "mode" TEXT NOT NULL DEFAULT 'missing',
-    "totalCount" INTEGER NOT NULL DEFAULT 0,
-    "queuedCount" INTEGER NOT NULL DEFAULT 0,
-    "runningCount" INTEGER NOT NULL DEFAULT 0,
-    "succeededCount" INTEGER NOT NULL DEFAULT 0,
-    "skippedCount" INTEGER NOT NULL DEFAULT 0,
-    "failedCount" INTEGER NOT NULL DEFAULT 0,
-    "maxAttempts" INTEGER NOT NULL DEFAULT 3,
-    "lastError" TEXT,
-    "startedAt" DATETIME,
-    "finishedAt" DATETIME,
-    "createdAt" DATETIME NOT NULL DEFAULT CURRENT_TIMESTAMP,
-    "updatedAt" DATETIME NOT NULL,
-    CONSTRAINT "AiBackfillBatch_tenantId_fkey" FOREIGN KEY ("tenantId") REFERENCES "Tenant" ("id") ON DELETE CASCADE ON UPDATE CASCADE,
-    CONSTRAINT "AiBackfillBatch_actorId_fkey" FOREIGN KEY ("actorId") REFERENCES "User" ("id") ON DELETE SET NULL ON UPDATE CASCADE
-);
-
--- CreateTable
-CREATE TABLE "AiBackfillItem" (
-    "id" TEXT NOT NULL PRIMARY KEY,
-    "tenantId" TEXT NOT NULL,
-    "batchId" TEXT NOT NULL,
-    "postId" TEXT NOT NULL,
-    "status" TEXT NOT NULL DEFAULT 'queued',
-    "attempts" INTEGER NOT NULL DEFAULT 0,
-    "nextRunAt" DATETIME NOT NULL DEFAULT CURRENT_TIMESTAMP,
-    "startedAt" DATETIME,
-    "finishedAt" DATETIME,
-    "lastError" TEXT,
-    "createdAt" DATETIME NOT NULL DEFAULT CURRENT_TIMESTAMP,
-    "updatedAt" DATETIME NOT NULL,
-    CONSTRAINT "AiBackfillItem_batchId_fkey" FOREIGN KEY ("batchId") REFERENCES "AiBackfillBatch" ("id") ON DELETE CASCADE ON UPDATE CASCADE,
-    CONSTRAINT "AiBackfillItem_postId_fkey" FOREIGN KEY ("postId") REFERENCES "Post" ("id") ON DELETE CASCADE ON UPDATE CASCADE
-);
-
--- CreateTable
-CREATE TABLE "AiBackfillLog" (
-    "id" TEXT NOT NULL PRIMARY KEY,
-    "tenantId" TEXT NOT NULL,
-    "batchId" TEXT NOT NULL,
-    "level" TEXT NOT NULL DEFAULT 'info',
-    "event" TEXT NOT NULL,
-    "message" TEXT NOT NULL,
-    "detail" JSONB NOT NULL DEFAULT '{}',
-    "createdAt" DATETIME NOT NULL DEFAULT CURRENT_TIMESTAMP,
-    CONSTRAINT "AiBackfillLog_batchId_fkey" FOREIGN KEY ("batchId") REFERENCES "AiBackfillBatch" ("id") ON DELETE CASCADE ON UPDATE CASCADE
-);
-
 -- CreateIndex
 CREATE UNIQUE INDEX "Tenant_slug_key" ON "Tenant"("slug");
 
@@ -681,40 +574,4 @@ CREATE INDEX "AuditLog_action_createdAt_idx" ON "AuditLog"("action", "createdAt"
 
 -- CreateIndex
 CREATE UNIQUE INDEX "TenantAiSettings_tenantId_key" ON "TenantAiSettings"("tenantId");
-
--- CreateIndex
-CREATE INDEX "SchoolEntity_tenantId_type_idx" ON "SchoolEntity"("tenantId", "type");
-
--- CreateIndex
-CREATE UNIQUE INDEX "SchoolEntity_tenantId_type_name_key" ON "SchoolEntity"("tenantId", "type", "name");
-
--- CreateIndex
-CREATE INDEX "SchoolModelSnapshot_tenantId_status_createdAt_idx" ON "SchoolModelSnapshot"("tenantId", "status", "createdAt");
-
--- CreateIndex
-CREATE UNIQUE INDEX "SchoolModelSnapshot_tenantId_version_key" ON "SchoolModelSnapshot"("tenantId", "version");
-
--- CreateIndex
-CREATE INDEX "PostAiAnalysis_tenantId_status_createdAt_idx" ON "PostAiAnalysis"("tenantId", "status", "createdAt");
-
--- CreateIndex
-CREATE UNIQUE INDEX "PostAiAnalysis_postId_key" ON "PostAiAnalysis"("postId");
-
--- CreateIndex
-CREATE INDEX "AiBackfillBatch_tenantId_status_createdAt_idx" ON "AiBackfillBatch"("tenantId", "status", "createdAt");
-
--- CreateIndex
-CREATE INDEX "AiBackfillItem_tenantId_status_nextRunAt_idx" ON "AiBackfillItem"("tenantId", "status", "nextRunAt");
-
--- CreateIndex
-CREATE INDEX "AiBackfillItem_batchId_status_nextRunAt_idx" ON "AiBackfillItem"("batchId", "status", "nextRunAt");
-
--- CreateIndex
-CREATE UNIQUE INDEX "AiBackfillItem_batchId_postId_key" ON "AiBackfillItem"("batchId", "postId");
-
--- CreateIndex
-CREATE INDEX "AiBackfillLog_tenantId_createdAt_idx" ON "AiBackfillLog"("tenantId", "createdAt");
-
--- CreateIndex
-CREATE INDEX "AiBackfillLog_batchId_createdAt_idx" ON "AiBackfillLog"("batchId", "createdAt");
